2015.11.21 merge github.py into getHosts.py
move to sub dir
